refactoring server/client sync stuff, started adding test case for that (wip)
