+++ src/client/components.rs
@@ -2,6 +2,3 @@ use bevy::prelude::Component;
 
 #[derive(Component)]
 pub struct Billboard;
-
-#[derive(Component)]
-pub struct Unsynced;
